src/requirements.txt
Django==5.2.5
djangorestframework
prometheus-client
requests
djangorestframework-simplejwt
python-decouple
celery==5.2.7
redis
kombu
tritonclient[all]
psycopg2-binary
grpcio
grpcio-tools
grpc-interceptor